requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.9
redis==5.0.8
requests==2.32.3
pydantic==2.8.2
apscheduler==3.10.4
boto3==1.34.162
python-jose==3.3.0
httpx==0.27.2
cachetools==5.4.0
python-dotenv==1.0.0
